requirements.txt
--allow-external pyPdf
--allow-unverified pyPdf
pyPdf
Django == 1.7.2
django-reversion == 1.8.5
python-dateutil >= 2.2, < 2.3
django-model-utils == 1.4.0
django-ajax-selects == 1.3.5
django-widget-tweaks < 1.2.0
django-tastypie == 0.12.1
interval == 1.0.0
python-ldap
reportlab >= 2.1, < 3.0
xhtml2pdf
https://codeload.github.com/ZTH1970/python-entrouvert/zip/master
django-localflavor
raven >= 3.5.2, < 3.6
M2Crypto
--allow-external pycairo
django_select2 < 4.3
django-journal
south
mysql-python
#PyGTK doesn't work with pip
#pdftk we need the binary
